HeadlessModeTest refractored
