+++ Browser_Actions/Browser_ActionTests.cs
@@ -5,6 +5,7 @@ using SeleniumExtras.WaitHelpers;
 using System.Drawing;
 using Functionality_Tests_Suit.Constants;
 using Functionality_Tests_Suit.FactoryPattern;
+using OpenQA.Selenium.Chrome;
 
 namespace Browser_Actions
 {
@@ -119,6 +120,25 @@ namespace Browser_Actions
             var currentSize = _driver.Manage().Window.Size;
             Assert.That(currentSize.Width, Is.EqualTo(newSize.Width), "Current window width is not 1000");
             Assert.That(currentSize.Height, Is.EqualTo(newSize.Height), "Current window height is not 800");
+        }
+
+        [Test]
+        public void HeadlessMode()
+        {
+            BrowserFactory.CloseDriver();
+            var options = new ChromeOptions();
+            options.AddArgument("--headless=new");
+            _driver = new ChromeDriver(options);
+
+            _driver.Navigate().GoToUrl(_mainUrl);
+            var elementCheckboxes = _driver.FindElement(By.CssSelector("[href*='checkboxes']"));
+            elementCheckboxes.Click();
+            var elementChecked = _driver.FindElement(By.XPath("//*[@id='checkboxes']/input[1]"));
+            elementChecked.Click();
+            Assert.That(elementChecked.Selected, Is.True, "checkbox 1 is not selected");
+            var elementUnchecked = _driver.FindElement(By.XPath("//*[@id='checkboxes']/input[2]"));
+            elementUnchecked.Click();
+            Assert.That(elementUnchecked.Selected, Is.False, "checkbox 2 is selected");
         }
 
         [OneTimeTearDown]
